pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "youtube-minder"
version = "0.1.0"
description = "Summarize YouTube videos via subtitles or audio transcription."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai==1.52.0",
    "python-dotenv==1.0.1",
    "streamlit>=1.34,<2",
    "yt-dlp==2025.11.12",
    "secretstorage==3.5.0",
    "httpx==0.27.2",
]

[dependency-groups]
dev = [
    "pytest==8.3.3",
]

[tool.hatch.build.targets.wheel]
packages = ["src/youtube_minder"]
